Move data field stages into core.datafields module
